requirements.txt
altgraph==0.17.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
fire==0.6.0
fonttools==4.53.1
importlib_resources==6.4.0
lxml==5.2.2
numpy==2.0.0
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.2
pefile==2023.2.7
pillow==10.4.0
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.1.2
pywin32-ctypes==0.2.2
setuptools==71.0.1
six==1.16.0
termcolor==2.4.0
typing_extensions==4.12.2
